Clean ups: More  Exceptions added - renaming classes
